working still on basic styling for the homepage
